Get working display mode toggle on home page
Uses a Svelte store to manage the state for the currently active display
modes across the app. The idea is that each page in the app will
subscribe to the active_display_modes store (although currently only the
home page does). When the value of that store is updated, the page
rerenders any of its contents which receive different CSS classes
depending on the display modes that are active.